DOC: fix examples in docstring

+++ pandas/core/indexes/base.py
@@ -1044,7 +1044,7 @@ class Index(IndexOpsMixin, PandasObject):
         >>> s.index.view(np.int64)
         array([-1,  0,  1])
         >>> s.index.view(np.float32)
-        array([nan, nan, 0.e+00, 0.e+00, 1.e-45, 0.e+00])
+        array([   nan,    nan, 0.e+00, 0.e+00, 1.e-45, 0.e+00], dtype=float32)
         >>> s.index.view(np.uint64)
         array([18446744073709551615,    0,  1])
         """
